Updated Server
Updated the main server Script. It now calls specific handlers to handle the different request sent by the photons or the VR application.

The handlers are set up so that they connect to the database instead of the server script to act like a proxy.

updated the script to find the suns position (will be updated later to be called by the oculus_handler)

updated the Db_helper so that the queries work more specifically for the request that will be made by the VR application

SO FAR: the server can run continueously and accept at most 5 connections at once. It can handle everything needed for the photon and just a basic login by the oculus app